Documentation and testing improvements.
Improved documentation of the webapp, webapp/auth and webapp/jwt
packages.

Updated test case organization so that when we implement more
jwt signing algorithms in the future, test cases can be easily
added.
