+++ jwt/doc.go
@@ -14,7 +14,7 @@
 
 /*
 
-Package jwt contains functions required for JWT signing and validation.
+Package jwt contains functions related to JWT signing and validation.
 
 Currently, only PS512 algorithm is supported, more methods will be added in
 future releases.
